1.1
* 1.0.2

* 1.0.1

This update makes the method CheckWinner() run a lot faster (probably not noticeable but hey the code is shorter AND uses less for loops). This update also prepares for a future feature of highlighting the chips that won the game.

* 1.0.1b

Updated file system

* 1.0.2

Added a dialog for restarting the game, changed "Reset Game" to "New Game", and changed the "No" option for restarting a new game to closing the dialog box rather than terminating the application.

* 1.0.2b

Edited whitespace and added a game.jar which is an executable file of the game!

* 1.0.2c

Coding is hard sometimes

* 1.1

Added glow to 3 of 4 possible win types.